[Abhishek] - Added RRSP

+++ app/src/main/java/com/aby/c0769778_w2020_mad3125_midterm/model/CRACustomer.java
@@ -8,7 +8,7 @@ import org.joda.time.LocalDate;
 public class CRACustomer implements Parcelable
 {
 
-    public CRACustomer(String SIN, String firstName, String lastName, Gender gender, LocalDate birthDate, Float grossIncome, Float rrspContributed) {
+    public CRACustomer(String SIN, String firstName, String lastName, Gender gender, LocalDate birthDate, Double grossIncome, Double rrspContributed) {
         this.SIN = SIN;
         this.firstName = firstName;
         this.lastName = lastName;
