+++ app/src/main/java/com/aby/c0769778_w2020_mad3125_midterm/ui/PersonInformationEntryActivity.java
@@ -186,8 +186,8 @@ public class PersonInformationEntryActivity extends AppCompatActivity {
                     edtLastNameText.getText().toString(),
                     getGender(),
                     HelperMethods.getInstance().stringToDate(edtDateText.getText().toString()),
-                    Float.parseFloat(edtGrossIncomeText.getText().toString()),
-                    Float.parseFloat(edtRRSPText.getText().toString()));
+                    Double.parseDouble(edtGrossIncomeText.getText().toString()),
+                    Double.parseDouble(edtRRSPText.getText().toString()));
             Intent mIntent = new Intent(PersonInformationEntryActivity.this, TaxDataDetailsActivity.class);
             mIntent.putExtra("CRACustomer", craCustomer);
             startActivity(mIntent);
